adds animation to addHost and addGuest imagebuttons in AddSelect.class
This adds bug where view is gone onBackPressed!
