+++ app/src/main/java/com/example/micha/chavrutamatch/AddSelect.java
@@ -1,11 +1,13 @@
 package com.example.micha.chavrutamatch;
 
+import android.animation.ObjectAnimator;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.view.ViewPropertyAnimatorCompatSet;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.TextView;
@@ -30,18 +32,45 @@ public class AddSelect extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.add_select);
         ButterKnife.bind(this);
+
     }
+//TODO fix bug on animate the imagebutton is gone on back pressed
     public void onAddGuestButtonClick(View view) {
-        Snackbar snackbar = Snackbar
-                .make(view, "To Open Chavurta Hosts-->", Snackbar.LENGTH_LONG);
-        snackbar.show();
-        Intent hostSelectIntent = new Intent(this, HostSelect.class);
-        startActivity(hostSelectIntent);
+        addHostButton.animate()
+                .alpha(0f)
+                .translationX(-addHostButton.getWidth())
+                .setDuration(getResources().getInteger(
+                        android.R.integer.config_shortAnimTime))
+                .withEndAction(new Runnable() {
+                    @Override
+                    public void run() {
+                        //end action ran after animation
+                        Intent hostSelectIntent = new Intent(AddSelect.this, HostSelect.class);
+                        startActivity(hostSelectIntent);
+                    }
+                });
+
     }
 
     public void onAddHostButtonClick(View view) {
-        Intent newHostIntent = new Intent(AddSelect.this, NewHost.class);
-        startActivity(newHostIntent);
+        //     WORKS   view.animate().alpha(0);
+        //   WORKS  view.animate().translationX(600).withLayer();
 
+        addHostButton.animate()
+                .alpha(0f)
+                .translationX(-addHostButton.getWidth())
+                .setDuration(getResources().getInteger(
+                        android.R.integer.config_shortAnimTime))
+                .withEndAction(new Runnable() {
+                    @Override
+                    public void run() {
+                        Intent newHostIntent = new Intent(AddSelect.this, NewHost.class);
+                        startActivity(newHostIntent);
+                    }
+                });
+    }
+
+    private void fadeAnimation(View v) {
+        v.animate().alpha(0);
     }
 }
