+++ app/src/main/java/com/example/micha/chavrutamatch/HostSelect.java
@@ -9,8 +9,10 @@ import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.view.ViewAnimationUtils;
+import android.view.ViewPropertyAnimator;
 import android.widget.EditText;
 import android.widget.FrameLayout;
+import android.widget.ImageButton;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -24,6 +26,8 @@ public class HostSelect extends AppCompatActivity {
     private OpenChavrutaAdapter mAdapter;
 @BindView(R.id.fl_host_pic)
     FrameLayout flHostPic;
+    @BindView(R.id.b_host_chavruta)
+            ImageButton ibHostChavruta;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -31,21 +35,7 @@ public class HostSelect extends AppCompatActivity {
         setContentView(R.layout.open_host_listview);
         ButterKnife.bind(this);
 
-        flHostPic.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-//                makeCircularRevealAnim(v);
-            }
-        });
-
 
     }
-    private void makeCircularRevealAnim(View v){
-        int finalRadius = (int) Math.hypot(v.getWidth()/2, v.getHeight()/2);
 
-        Animator anim = ViewAnimationUtils.createCircularReveal(
-                v, (int) v.getWidth()/2, (int) v.getHeight()/2, 0, finalRadius);
-        v.setBackgroundColor(Color.GREEN);
-        anim.start();
-    }
 }
